+++ tests/chess_dataset_test.py
@@ -243,37 +243,37 @@ def test_chess_dataset():
         # Total = 21
         assert tc['data'].shape == (21, 8, 8)
 
-        assert tc['data'][0].equal(torch.from_numpy(tc['curr_p']))
-        assert tc['data'][1].equal(torch.from_numpy(tc['curr_n']))
-        assert tc['data'][2].equal(torch.from_numpy(tc['curr_b']))
-        assert tc['data'][3].equal(torch.from_numpy(tc['curr_r']))
-        assert tc['data'][4].equal(torch.from_numpy(tc['curr_q']))
-        assert tc['data'][5].equal(torch.from_numpy(tc['curr_k']))
+        assert tc['data'][0].equal(torch.from_numpy(tc['curr_p']).float())
+        assert tc['data'][1].equal(torch.from_numpy(tc['curr_n']).float())
+        assert tc['data'][2].equal(torch.from_numpy(tc['curr_b']).float())
+        assert tc['data'][3].equal(torch.from_numpy(tc['curr_r']).float())
+        assert tc['data'][4].equal(torch.from_numpy(tc['curr_q']).float())
+        assert tc['data'][5].equal(torch.from_numpy(tc['curr_k']).float())
 
-        assert tc['data'][6].equal(torch.from_numpy(tc['oppo_p']))
-        assert tc['data'][7].equal(torch.from_numpy(tc['oppo_n']))
-        assert tc['data'][8].equal(torch.from_numpy(tc['oppo_b']))
-        assert tc['data'][9].equal(torch.from_numpy(tc['oppo_r']))
-        assert tc['data'][10].equal(torch.from_numpy(tc['oppo_q']))
-        assert tc['data'][11].equal(torch.from_numpy(tc['oppo_k']))
+        assert tc['data'][6].equal(torch.from_numpy(tc['oppo_p']).float())
+        assert tc['data'][7].equal(torch.from_numpy(tc['oppo_n']).float())
+        assert tc['data'][8].equal(torch.from_numpy(tc['oppo_b']).float())
+        assert tc['data'][9].equal(torch.from_numpy(tc['oppo_r']).float())
+        assert tc['data'][10].equal(torch.from_numpy(tc['oppo_q']).float())
+        assert tc['data'][11].equal(torch.from_numpy(tc['oppo_k']).float())
 
         assert tc['data'][12].equal(
-            torch.from_numpy(np.full(BOARD_SIZE, tc['rep_2'])))
+            torch.from_numpy(np.full(BOARD_SIZE, tc['rep_2'])).float())
         assert tc['data'][13].equal(
-            torch.from_numpy(np.full(BOARD_SIZE, tc['rep_3'])))
+            torch.from_numpy(np.full(BOARD_SIZE, tc['rep_3'])).float())
         assert tc['data'][14].equal(
-            torch.from_numpy(np.full(BOARD_SIZE, tc['turn'])))
+            torch.from_numpy(np.full(BOARD_SIZE, tc['turn'])).float())
         assert tc['data'][15].equal(
-            torch.from_numpy(np.full(BOARD_SIZE, tc['move_count'])))
+            torch.from_numpy(np.full(BOARD_SIZE, tc['move_count'])).float())
         assert tc['data'][16].equal(
-            torch.from_numpy(np.full(BOARD_SIZE, tc['b_k_castling'])))
+            torch.from_numpy(np.full(BOARD_SIZE, tc['b_k_castling'])).float())
         assert tc['data'][17].equal(
-            torch.from_numpy(np.full(BOARD_SIZE, tc['b_q_castling'])))
+            torch.from_numpy(np.full(BOARD_SIZE, tc['b_q_castling'])).float())
         assert tc['data'][18].equal(
-            torch.from_numpy(np.full(BOARD_SIZE, tc['w_k_castling'])))
+            torch.from_numpy(np.full(BOARD_SIZE, tc['w_k_castling'])).float())
         assert tc['data'][19].equal(
-            torch.from_numpy(np.full(BOARD_SIZE, tc['w_q_castling'])))
+            torch.from_numpy(np.full(BOARD_SIZE, tc['w_q_castling'])).float())
         assert tc['data'][20].equal(
-            torch.from_numpy(np.full(BOARD_SIZE, tc['no_progress'])))
+            torch.from_numpy(np.full(BOARD_SIZE, tc['no_progress'])).float())
 
         assert tc['move'].equal(torch.Tensor([tc['expected_move']]))
